pyproject.toml
[tool.poetry]
name = "ucllm-nedo-prod"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11, <4.0"
hojichar = "^0.9.0"
fugashi = "^1.3.1"
gensim = "^4.3.2"
numpy = "^1.26.4"
polars = {extras = ["all"], version = "^0.20.21"}
aiohttp = "^3.9.5"
pyyaml = "^6.0.1"

[tool.poetry.group.dev.dependencies]
black = "^24.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
